Move the project to a Rails-like folder structure
Maintaining a React project in a single file is not scalable.
There are different options for organising the folder structure,
such as the chosen one, inspired on Ruby on Rails, where folders
are organised by their functions: actions, componentes, reducers.

Another option would be a featured-based folder organisation, where
folders are named after each component or feature, and each folder
would have a file exporting their component, action, reducer and
so on.